rewrite documentation addition renamed folders accordingly
