+++ src/react-component-features/disposable-decorator.ts
