+++ src/react-component-features/properties-feature.ts
