+++ test/react-component-features/disposable-decorator.spec.tsx
