+++ test/react-component-features/properties-feature.spec.tsx
